requirements.txt
transformers>=4.30.0
datasets==2.14.7
torch>=2.0.0
pyarrow>=13,<16
evaluate>=0.4.1
rouge-score>=0.1.2
pandas>=2.0.0
numpy<2.0
pyyaml>=6.0.1
matplotlib>=3.8.0
seaborn>=0.13.0
tqdm>=4.66.0
streamlit>=1.32.0
bert_score
